+++ packages/at_secondary_server/lib/src/verb/handler/sync_progressive_verb_handler.dart
@@ -40,13 +40,25 @@ class SyncProgressiveVerbHandler extends AbstractVerbHandler {
     // Get Commit Log Instance.
     var atCommitLog = await (AtCommitLogManagerImpl.getInstance()
         .getCommitLog(AtSecondaryServerImpl.getInstance().currentAtSign));
+    int? skipDeletesUntil = verbParams[AtConstants.skipDeletesUntil] != null
+        ? int.parse(verbParams[AtConstants.skipDeletesUntil]!)
+        : null;
+    int? syncLimit = verbParams[AtConstants.syncLimit] != null
+        ? int.parse(verbParams[AtConstants.syncLimit]!)
+        : null;
     // Get entries to sync
-    var commitEntryIterator = atCommitLog!.getEntries(
+    Iterator<MapEntry<String, CommitEntry>> commitEntryIterator;
+    // if client doesn't pass syncLimit set the default value from server
+    syncLimit ??= AtSecondaryConfig.syncPageLimit;
+    commitEntryIterator = atCommitLog!.getEntries(
         int.parse(verbParams[AtConstants.fromCommitSequence]!) + 1,
-        regex: verbParams['regex']);
+        regex: verbParams['regex'],
+        skipDeletesUntil: skipDeletesUntil,
+        limit: syncLimit);
 
     List<KeyStoreEntry> syncResponse = [];
-    await prepareResponse(capacity, syncResponse, commitEntryIterator,
+    await prepareResponse(
+        capacity, syncLimit, syncResponse, commitEntryIterator,
         enrollmentId:
             (atConnection.metaData as InboundConnectionMetadata).enrollmentId);
 
@@ -57,8 +69,11 @@ class SyncProgressiveVerbHandler extends AbstractVerbHandler {
   /// 1. there is at least one item in [syncResponse], and the response length is greater than [desiredMaxSyncResponseLength], or
   /// 2. there are [AtSecondaryConfig.syncPageLimit] items in the [syncResponse]
   @visibleForTesting
-  Future<void> prepareResponse(int desiredMaxSyncResponseLength,
-      List<KeyStoreEntry> syncResponse, Iterator<dynamic> commitEntryIterator,
+  Future<void> prepareResponse(
+      int desiredMaxSyncResponseLength,
+      int syncPageLimit,
+      List<KeyStoreEntry> syncResponse,
+      Iterator<dynamic> commitEntryIterator,
       {String? enrollmentId}) async {
     int currentResponseLength = 0;
     Map<String, String> enrolledNamespaces = {};
@@ -69,9 +84,8 @@ class SyncProgressiveVerbHandler extends AbstractVerbHandler {
               .get(enrollmentId))
           .namespaces;
     }
-
-    while (commitEntryIterator.moveNext() &&
-        syncResponse.length < AtSecondaryConfig.syncPageLimit) {
+    while (
+        commitEntryIterator.moveNext() && syncResponse.length < syncPageLimit) {
       var atKeyType = AtKey.getKeyType(commitEntryIterator.current.key,
           enforceNameSpace: false);
       if (atKeyType == KeyType.invalidKey) {
@@ -150,11 +164,24 @@ class SyncProgressiveVerbHandler extends AbstractVerbHandler {
     if (metaData == null) {
       return metaDataMap;
     }
-    metaData.toJson().forEach((key, value) {
-      if (value != null) {
-        metaDataMap[key] = value.toString();
+    Iterator itr = metaData.toJson().entries.iterator;
+    while (itr.moveNext()) {
+      // The value of [AtConstants.sharedWithPublicKeyHash] stores a Map containing
+      // the hash value and the hashing algorithm used for hashing the data.
+      // For example, {"hash":"dummy_value", "hashingAlgo":"sha512"}.
+      // Using toString() will not allow convert this into a Map, which is necessary
+      // for constructing the PublicKeyHash type on the client side.
+      // Therefore, a JSON-encoded string is used here, and on the client side,
+      // "jsonDecode" will be used to retrieve the Map and build the PublicKeyHash instance.
+      if (itr.current.key == AtConstants.sharedWithPublicKeyHash &&
+          itr.current.value != null) {
+        metaDataMap[itr.current.key] = jsonEncode(itr.current.value);
+        continue;
       }
-    });
+      if (itr.current.value != null) {
+        metaDataMap[itr.current.key] = itr.current.value.toString();
+      }
+    }
     return metaDataMap;
   }