+++ packages/at_persistence_secondary_server/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 3.1.0
+- feat: commit log changes for sync skipDeletesUntil feature
+- build[deps]: Upgraded the following package:
+  - at_commons to v5.1.0
+## 3.0.66
+- feat: Add "PublicKeyHash" to the "AtMetadata" which holds the hash value of encryption public key
+- build[deps]: Upgraded the following packages:
+  - at_commons to v5.0.2
+  - lints to v5.0.0
+  - test to v1.25.8
 ## 3.0.65
 - fix: Modified checks in commit log keystore _alwaysIncludeInSync method to match only reserved shared_key,
   encryption public key and public key without namespace.